Comments
Overall clear and understandable. Below are some small improvement suggestions.

"Methodology and prototype software to combine these different data sources, 
and to retrieve target classification and other products, were developed within 
the so-called Cloudnet project [@IllingworthEtAl07]."
Confusing sentence, maybe split to two or more? Something like "The so-called Cloudnet project was formed to unify processing standards and develop prototype software (or whatever was its purpose). Due to the project, it is now possible to combine these data..."

I didn't understand some of the terms, such as scattering objects, but if they are as clear as the day for users it's not a problem. Do you think an exhaustive list of improvements compared to original cloudnet (e.g. changelog) would be useful?