Summary: Changes to mira2nc to support znc input (and STSR) for the whole of cloudnet
Added 2 helper functions to mira2nc which handle the specific keymap for the filetype (mmclx or znc) and removal of DropSize which according to the MetekManual has the dimensions range and Doppler (but no time as it is assumed to change little). This creates a concat conlfict and thus requires removal. Minor error catching for empty filelist as the recent change to be able to pass in a list requires a filetype and may lead to a potentially empty filelist.

Closes #83